arclamp-log: abort if no message received in 30 minutes
Make arclamp-log throw an exception (and exit) if 30m elapse where no
message was received. This ensures that arclamp-log is restarted if the
connection to redis is interrupted. The timeout is long (30m) to avoid
crash-looping / log spam on the beta cluster, where profiler samples are
infrequent. A half-hour gap in arclamp logs is not a big deal.

I didn't make this configurable because it's a bit of a kludge. It would
be better to use health-checks instead[1], but the version of redis-py
on WMF servers is too old. So I propose going with this approach for
now, and revisiting it when redis-py is updated.

[1]: https://github.com/andymccurdy/redis-py/commit/d9227fe40c

Tested on beta.

Bug: T215740
Change-Id: Ie0d853685237f89cc4a3077a49ae04026cb230e7