Support reading gzip'd logs
This adds support for reading gzip'd log files in arclamp-grep.py and
arclamp-generate-svgs, and for expiring gzip'd log files in
arclamp-log.py.

Using any compression format other than gzip (e.g. bzip2 or xz) will
cause data loss and failure to expire old logfiles.  (The first
version of this patch did not have this limitation, but reviewers
objected to making our tools more robust.)

I've also changed the tools to use mtime instead of ctime; this is
needed for correct operation if we ever move the log files or restore
them from backup.  Because mtime is easy to change, this will also let
us keep the original log timestamp when we compress a file after the
fact.

Bug: T235456
Change-Id: I1e7802c01e1ae45bd39675462c9030c45b06e35e