Add metrics for arclamp-compress-logs
Compressed vs. non-compressed logs are labeled when reporting oldest
and newest input files, as is the timestamp of the oldest temporary
file (if any).  This can be used to monitor if arclamp-compress-logs
is working as intended.

Per feedback on the previous incarnation of this script, I've added
code to ensure old temporary files are cleaned up, as well as
protection against bogus filenames corrupting the output.  The
"instance" label is now "profiler", to fix a naming collision with
Prometheus's automatic labels.

Bug: T256035
Change-Id: Id53c120111ec2f6b25721e4530f288009600104d