Prevent parallel execution and check .svgz size
We're running out of memory in production when running flamegraph.pl
on large files.  This change tries to mitigate the issue by only
allowing one copy of arclamp-generate-svgs to be running at a time.

By removing parallelism, this risks the pipeline stalling if it keeps
re-attempting the same file over and over again.  To partially protect
against this, we now shuffle the processing order.  We also now
correctly handle filenames containing whitespace.

When an OOM is encountered, flamegraph.pl leaves around 20 byte files
containing just a gzip header.  The script now detects those and
attempts to regenerate them.  (It already ignored zero-length files.)

I've also changed the script to no longer "nice" its subprocesses.
Generating these graphs is the primary function of webperfX002, so
let's prioritize appropriately.

I'm leaving in place the parallel generation of forward and reverse
files, added in I540e4adf.  If we're still OOM'ing after this patch
lands, we can change it to also process those serially.  (If that also
fails, the next step will be to request a Ganeti instance with a
couple of additional dollars worth of RAM.)

Bug: T259167
Change-Id: I247b774d45d041984fe674ba34fa27b09d451ef2