Process large files serially and improve metrics
If the input file is larger than 10e8 bytes, don't try to generate
forward and reverse flamegraphs in parallel.  This burns some extra
CPU but should help avoid out-of-memory errors.

I've also updated arclamp-generate-metrics to enable alerting on the
failure mode being addressed here.  When reporting the timestamp of
the newest SVG file, we now ignore files which are 20 bytes or
smaller, as we do when figuring out which files need to be
re-generated.  The count of small (truncated) files is now reported in
its own metric.

This patch also fixes a problem which prevented effective alerting on
stale temporary files from arclamp-compress-logs.

Bug: T259167
Change-Id: Ic697ad6befc6cce7c4b9fe5cec21c1c1baba16c4