Skip flamegraph execution if input is empty
flamegraph.pl outputs "ERROR: No stack counts found" on stderr when
the input is empty.  This happens most often at the top of the hour,
when we don't yet have sufficient data to generate the SVG, and is
much more frequent in beta due to lower traffic.  It can also happen
if the process feeding stdin is killed for some reason, e.g. OOM.

This patch uses the ifne command from moreutils to only launch
flamegraph.pl if stdin is non-empty.

The moreutils package is already installed on WMF hosts; see
modules/base/manifests/standard_packages.pp in operations/puppet.

Change-Id: Ib1eec01ef2abd3a6c17e01a5bd8910e0350b800e