Process serially above 100MB, not 1GB
I set the threshold too low when coming up with the fix in Ic697ad6b.
Most daily files need to be processed serially to avoid flamegraph.pl
running out of memory, even with 16GB of RAM.

The other mitigations contemplated in the linked task have less
urgency so will be implemented later.

Bug: T259167
Change-Id: Ibad670e9fbd2115d130e1835b6d1dbfb611ec40a