Remove mention of outdated /srv/xenon
All scripts in this repo use /srv/arclamp and in WMF production,
this is symlinked from /srv/xenon for compat. As far as I know,
arclamp-grep can follow symlinks.

Also fix the copyright year of the original version to be 2014
instead of 2015. It was merged in git in Jan 2015 but developed
in 2014. This isn't a big deal, but the reason I'm fixing it is
that the source tree under the 1.0.0 tag actually says
"Copyright 2014", which we link to from the README after saying
it is created in 2015 – confusing indeed.

Change-Id: Ib1aa8c2624554a7356da79489ed352d956cd66cf